% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{aff_clust_mean}
\alias{aff_clust_mean}
\title{Find mean affinity from each element to each cluster}
\usage{
aff_clust_mean(sim_mat, cast_ob, aff_thres)
}
\description{
Find mean affinity from each element to each cluster
}
