% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{aff_cluster_between}
\alias{aff_cluster_between}
\title{Across cluster affinity
An assymetric distance, where the distance from a to b
is the average affinity of elements in a to elements in b
returns a long form dataframe}
\usage{
aff_cluster_between(cast_obj, sim_mat, aff_thres)
}
\description{
Across cluster affinity
An assymetric distance, where the distance from a to b
is the average affinity of elements in a to elements in b
returns a long form dataframe
}
