% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_cautious.R
\name{cast_alg_cautious}
\alias{cast_alg_cautious}
\title{Two functions that extend cast to reduce the
number of clusters}
\usage{
cast_alg_cautious(sim_mat, aff_thres, max_iter = 20)
}
\description{
cast_alg_cautious stabilizes the matrix after every cluster is added,
so we don't create a lot of small clusters on the edges.
I expect to use all the sites faster.
}
