% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_compact.R
\name{cast_compact}
\alias{cast_compact}
\title{cast_compact}
\usage{
cast_compact(cast_ob, sim_mat, aff_thres, max_iter = nrow(sim_mat) * 2)
}
\description{
takes a stabilized cast object, and deletes
one cluster, assignes the sites to nearest clusters,
stabilizes again, and
checks whether the average cluster affinity
remains above the threshold. If the threshold is still valid,
keep the changed cast algorithm, otherwise revert and try another cluster.
}
