% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_heirarchy.R
\name{cast_h}
\alias{cast_h}
\title{heirarchical grouping of cast objects
the aff_thres_vec will be sorted
this function is recursive}
\usage{
cast_h(
  sim_mat,
  aff_thres_vec,
  stabilize = FALSE,
  max_stab_iter = nrow(sim_mat) * 4
)
}
\description{
heirarchical grouping of cast objects
the aff_thres_vec will be sorted
this function is recursive
}
