% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_heirarchy.R
\name{cast_h_reorder}
\alias{cast_h_reorder}
\title{unrolls the cast clusters to the desired depth
also recursive}
\usage{
cast_h_reorder(cast_h_obj, depth = length(cast_h_obj), reordering = NULL)
}
\description{
unrolls the cast clusters to the desired depth
also recursive
}
