% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_k.R
\name{cast_k}
\alias{cast_k}
\title{CAST algorithm for a given number of clusters.}
\usage{
cast_k(sim_mat, k, aff_thres = mean(sim_mat), max_iter = nrow(sim_mat) * 2)
}
\arguments{
\item{sim_mat}{symmetric matrix with values in the range [0,1]}

\item{k}{number of clusters to fit}

\item{max_iter}{don't shuffle objects forever}
}
\description{
Rather than fixing the threshold and
finding k that meets the threshold,
this function fixes k regardless of
threshold.
}
\details{
I tried just allocating all the seeds at once
but it tends to find poor local optima
}
