% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_obj_to_df.R
\name{cast_obj_to_df}
\alias{cast_obj_to_df}
\title{Convert CAST object to a data.frame mapping elements to clusters}
\usage{
cast_obj_to_df(cast_ob)
}
\arguments{
\item{cast_ob}{A CAST object returned by a CAST method}
}
\value{
data.frame with $elem and $clust cols. Elem is a row id of the data passed into the CAST method.
}
\description{
CAST objects are implemented as a list of clusters, with each
cluster storing the set of elements making up the cluster.
Plotting usually needs a data.frame. This helper function
produces a ggplot ready data.frame.
}
