% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_optimal.R
\name{cast_optimal_recurse}
\alias{cast_optimal_recurse}
\title{Recursive calculation of optimal aff_thres}
\usage{
cast_optimal_recurse(
  sim_mat,
  aff_range,
  m,
  min_range,
  min_tol,
  rec_data,
  rec_depth
)
}
\description{
`cast_opminal` is a wrapper function that
hides recursive parameters from the user.
}
