% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_stabilize.R
\name{cast_stabilize}
\alias{cast_stabilize}
\title{Stabilize cluster membership}
\usage{
cast_stabilize(cast_obj, aff_thres, sim_mat, max_iter = nrow(sim_mat) * 2)
}
\description{
cast_stabilize() works iteratively, updating one site at a time
and recalculating the affinities each time.
cast_stabilize() will only make an update if it either maintains
all the average within cluster affinities above the threshold, or
increases the within cluster affinities if one is already below the
threshold.
}
