% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_stabilize.R
\name{cast_stabilize_batch}
\alias{cast_stabilize_batch}
\title{Stabilize cluster membership}
\usage{
cast_stabilize_batch(cast_obj, aff_thres, sim_mat, max_iter = 20)
}
\description{
cast_stabilize_batch() does a bulk update, finding all sites that
could be assigned to another cluster before recalulating the affinities.
cast_stabilize_batch() always proceeds without checking the impact
the update has on within-cluster affinities.
}
