% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{.aff_clust_mean}
\alias{.aff_clust_mean}
\title{Internal function for aff_clust_mean,}
\usage{
.aff_clust_mean(sim_mat, cast_ob, aff_thres)
}
\description{
This function may be called directly if you
can ensure that the diagonals of sim_mat
are set to NA  `diag(sim_mat) <- NA`
to speed up computation.
}
