% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_plot.R
\name{gg_sim_mat}
\alias{gg_sim_mat}
\title{ggplot object of the similarity matrix}
\usage{
gg_sim_mat(
  sim_mat,
  cast_ob = NULL,
  sort_among_clust = TRUE,
  sort_within_clust = TRUE,
  highlight = FALSE,
  legend_label = "Similarity"
)
}
\description{
Without a cast object, it will just plot them
similarity matrix as is, with no particular order.
}
\details{
With a cast object, the similarity matrix will
be reordered to put sites in the same cluster
together.

If secondary_sort is true, then the clusters will
be further reordered to put similar clusters together,
using a dendrogram sort.

if highlight = TRUE, then rectangles will be drawn
around each cluster.
}
