% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{membership_mat}
\alias{membership_mat}
\title{Calculate membership matix from cast object}
\usage{
membership_mat(cast_ob, is_long = FALSE)
}
\description{
Helper function for hubert_gamma(),
}
\details{
Produces a matrix where element (i,j)
is 1 if site i and site j are in the same
cluster, otherwise 0.

is_long toggles between long form and wide form
matrix for return values.
For very large areas, a full nxn matrix may
not fit it memory. However, the matrix becomes
sparse, as k increases, so long form should always work.
}
