% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_plot.R
\name{sort_between}
\alias{sort_between}
\title{Get sort order of clusters, so similar clusters are put together}
\usage{
sort_between(sim_mat, cast_ob)
}
\description{
Get sort order of clusters, so similar clusters are put together
}
